LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/compile-error-data-url-expected.txt

CONSOLE MESSAGE: line 1: SyntaxError: Unexpected token ')'

FAIL window.onerror - compile error in <script src=data:...> assert_equals: second arg expected "data:text/javascript,for(;){}" but got ""
PASS window.onerror - compile error in <script src=data:...> (column) 
